Arrange buttons
It was difficult to get the last row middle button to grow at the exact
rate of the other square buttons, while also keeping a uniform gap. This
was solved by calculating exactly what percentage-width the square buttons
must be based on the gap width and the number of gaps. This way, all the
square buttons can be set to not have any growing behavior, and the two
rectangular buttons on the bottom can have all the growing to
themselves.