Enable numbers, CE button, and digital display
CE stands for 'clear entry' and clears only the current number being entered.
The C button will clear both the current number and any previous number
and operator input as well, but it is not yet implemented.

The current number `displayValue` is stored as a string, so that decimal
points will work correctly later.