Remove strict viewport units to allow zoom
This change takes an average between the old viewport units and a
rem-based unit that looks similar at default zoom on my computer. It is
now possible to zoom in, but there may be some screens where the
default, un-zoomed state is for the calculator to not fit on the page.

I think it's worth it to make this change because it was broken
accessibility-wise before, but it's still disappointing. Ideally, the
calculator would start out entirely in-screen, but zoom would still
behave as expected. I do not know how to do this without some sort of
metric that exposed zoom to the CSS. Maybe I need to be doing something
clever with min() or max()?

The viewport units accessibility problem is documented in [a WCAG
failure entry](https://www.w3.org/WAI/WCAG21/Techniques/failures/F94),
but their only solution is to use a different unit or a JavaScript
control for zooming text. Others have brought up solutions using a
combination of units like [Chip
Cullen](https://chipcullen.com/dont-use-viewport-units-for-font-size-on-their-own/)
and [Amelia Bellamy-Royds](https://codepen.io/AmeliaBR/pen/RNXeyd), but
I couldn't find anyone who's already solved this issue yet with the amount
of control that I'd want.